auditLog table and corrections

+++ src/main/java/com/example/OCBReporting/repository/FieldRepository.java
@@ -4,5 +4,5 @@ import com.example.OCBReporting.model.App;
 import com.example.OCBReporting.model.Field;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FieldRepository extends JpaRepository<Field,Integer> {
+public interface FieldRepository extends JpaRepository<Field,Long> {
 }
