+++ src/main/java/com/example/OCBReporting/repository/TableRepository.java
@@ -4,5 +4,5 @@ import com.example.OCBReporting.model.Query;
 import com.example.OCBReporting.model.Table;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TableRepository extends JpaRepository<Table,Integer> {
+public interface TableRepository extends JpaRepository<Table,Long> {
 }
